US2215773: move client, builder, impl, disposer to root folder; add new methods
