+++ access-checkout/src/main/java/com/worldpay/access/checkout/client/AccessCheckoutClientImpl.kt
@@ -1,14 +1,14 @@
-package com.worldpay.access.checkout.session
+package com.worldpay.access.checkout.client
 
 import android.content.Context
 import android.content.Intent
-import com.worldpay.access.checkout.client.session.AccessCheckoutClient
 import com.worldpay.access.checkout.client.session.model.CardDetails
 import com.worldpay.access.checkout.client.session.model.SessionType
+import com.worldpay.access.checkout.session.ActivityLifecycleObserver
+import com.worldpay.access.checkout.session.ActivityLifecycleObserverInitialiser
 import com.worldpay.access.checkout.session.broadcast.LocalBroadcastManagerFactory
 import com.worldpay.access.checkout.session.broadcast.receivers.NUM_OF_SESSION_TYPES_REQUESTED
 import com.worldpay.access.checkout.session.broadcast.receivers.SessionBroadcastReceiver
-import com.worldpay.access.checkout.session.broadcast.receivers.SessionBroadcastReceiver.Companion.NUMBER_OF_SESSION_TYPE_KEY
 import com.worldpay.access.checkout.session.handlers.SessionRequestHandlerFactory
 
 internal class AccessCheckoutClientImpl(
@@ -34,7 +34,7 @@ internal class AccessCheckoutClientImpl(
 
     private fun broadcastSessionTypeInfo(sessionTypes: List<SessionType>) {
         val broadcastIntent = Intent(context, SessionBroadcastReceiver::class.java)
-        broadcastIntent.putExtra(NUMBER_OF_SESSION_TYPE_KEY, sessionTypes.size)
+        broadcastIntent.putExtra(SessionBroadcastReceiver.Companion.NUMBER_OF_SESSION_TYPE_KEY, sessionTypes.size)
         broadcastIntent.action = NUM_OF_SESSION_TYPES_REQUESTED
         localBroadcastManagerFactory.createInstance().sendBroadcast(broadcastIntent)
     }
