+++ access-checkout/src/main/java/com/worldpay/access/checkout/session/AccessCheckoutClientDisposer.kt
@@ -1,6 +1,7 @@
 package com.worldpay.access.checkout.session
 
-import com.worldpay.access.checkout.client.session.AccessCheckoutClient
+import com.worldpay.access.checkout.client.AccessCheckoutClient
+import com.worldpay.access.checkout.client.AccessCheckoutClientImpl
 
 open class AccessCheckoutClientDisposer {
     open fun dispose(accessCheckoutClient: AccessCheckoutClient) {
